move Mode enum to _models.py
